+++ pkgs/web_socket_channel/lib/io.dart
@@ -32,8 +32,15 @@ class IOWebSocketChannel extends StreamChannelMixin
 
   @override
   final Stream stream;
+
   @override
   final WebSocketSink sink;
+
+  /// Completer for [ready].
+  final Completer<void> _readyCompleter;
+
+  @override
+  Future<void> get ready => _readyCompleter.future;
 
   /// The underlying [WebSocket], if this channel has connected.
   ///
@@ -55,6 +62,10 @@ class IOWebSocketChannel extends StreamChannelMixin
   /// [pingInterval]. It defaults to `null`, indicating that ping messages are
   /// disabled.
   ///
+  /// [connectTimeout] determines how long to wait for [WebSocket.connect]
+  /// before throwing a [TimeoutException]. If connectTimeout is null then the
+  /// connection process will never time-out.
+  ///
   /// If there's an error connecting, the channel's stream emits a
   /// [WebSocketChannelException] wrapping that error and then closes.
   factory IOWebSocketChannel.connect(
@@ -62,21 +73,28 @@ class IOWebSocketChannel extends StreamChannelMixin
     Iterable<String>? protocols,
     Map<String, dynamic>? headers,
     Duration? pingInterval,
+    Duration? connectTimeout,
   }) {
     late IOWebSocketChannel channel;
     final sinkCompleter = WebSocketSinkCompleter();
-    final stream = StreamCompleter.fromFuture(
-      WebSocket.connect(url.toString(), headers: headers, protocols: protocols)
-          .then((webSocket) {
-        webSocket.pingInterval = pingInterval;
-        channel._webSocket = webSocket;
-        sinkCompleter.setDestinationSink(_IOWebSocketSink(webSocket));
-        return webSocket;
-      }).catchError(
-        (Object error) => throw WebSocketChannelException.from(error),
-      ),
+    var future = WebSocket.connect(
+      url.toString(),
+      headers: headers,
+      protocols: protocols,
     );
-
+    if (connectTimeout != null) {
+      future = future.timeout(connectTimeout);
+    }
+    final stream = StreamCompleter.fromFuture(future.then((webSocket) {
+      webSocket.pingInterval = pingInterval;
+      channel._webSocket = webSocket;
+      channel._readyCompleter.complete();
+      sinkCompleter.setDestinationSink(_IOWebSocketSink(webSocket));
+      return webSocket;
+    }).catchError((Object error, StackTrace stackTrace) {
+      channel._readyCompleter.completeError(error, stackTrace);
+      throw WebSocketChannelException.from(error);
+    }));
     return channel =
         IOWebSocketChannel._withoutSocket(stream, sinkCompleter.sink);
   }
@@ -86,7 +104,8 @@ class IOWebSocketChannel extends StreamChannelMixin
       : _webSocket = socket,
         stream = socket.handleError(
             (error) => throw WebSocketChannelException.from(error)),
-        sink = _IOWebSocketSink(socket);
+        sink = _IOWebSocketSink(socket),
+        _readyCompleter = Completer()..complete();
 
   /// Creates a channel without a socket.
   ///
@@ -95,7 +114,8 @@ class IOWebSocketChannel extends StreamChannelMixin
   IOWebSocketChannel._withoutSocket(Stream stream, this.sink)
       : _webSocket = null,
         stream = stream.handleError(
-            (error) => throw WebSocketChannelException.from(error));
+            (error) => throw WebSocketChannelException.from(error)),
+        _readyCompleter = Completer();
 }
 
 /// A [WebSocketSink] that forwards [close] calls to a `dart:io` [WebSocket].