+++ pkgs/web_socket_channel/test/io_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('vm')
+import 'dart:async';
 import 'dart:io';
 
 import 'package:test/test.dart';
@@ -28,6 +29,8 @@ void main() {
     final webSocket = await WebSocket.connect('ws://localhost:${server.port}');
     final channel = IOWebSocketChannel(webSocket);
 
+    expect(channel.ready, completes);
+
     var n = 0;
     channel.stream.listen((message) {
       if (n == 0) {
@@ -47,6 +50,7 @@ void main() {
 
   test('.connect communicates immediately', () async {
     server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
     server.transform(WebSocketTransformer()).listen((WebSocket webSocket) {
       final channel = IOWebSocketChannel(webSocket);
       channel.stream.listen((request) {
@@ -56,6 +60,9 @@ void main() {
     });
 
     final channel = IOWebSocketChannel.connect('ws://localhost:${server.port}');
+
+    expect(channel.ready, completes);
+
     channel.sink.add('ping');
 
     channel.stream.listen(
@@ -69,6 +76,7 @@ void main() {
   test('.connect communicates immediately using platform independent api',
       () async {
     server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
     server.transform(WebSocketTransformer()).listen((WebSocket webSocket) {
       final channel = IOWebSocketChannel(webSocket);
       channel.stream.listen((request) {
@@ -79,6 +87,9 @@ void main() {
 
     final channel =
         WebSocketChannel.connect(Uri.parse('ws://localhost:${server.port}'));
+
+    expect(channel.ready, completes);
+
     channel.sink.add('ping');
 
     channel.stream.listen(
@@ -91,6 +102,7 @@ void main() {
 
   test('.connect with an immediate call to close', () async {
     server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
     server.transform(WebSocketTransformer()).listen((WebSocket webSocket) {
       expect(() async {
         final channel = IOWebSocketChannel(webSocket);
@@ -101,18 +113,23 @@ void main() {
     });
 
     final channel = IOWebSocketChannel.connect('ws://localhost:${server.port}');
+
+    expect(channel.ready, completes);
+
     await channel.sink.close(5678, 'raisin');
   });
 
   test('.connect wraps a connection error in WebSocketChannelException',
       () async {
     server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
     server.listen((request) {
       request.response.statusCode = 404;
       request.response.close();
     });
 
     final channel = IOWebSocketChannel.connect('ws://localhost:${server.port}');
+    expect(channel.ready, throwsA(isA<WebSocketException>()));
     expect(channel.stream.drain(), throwsA(isA<WebSocketChannelException>()));
   });
 
@@ -122,6 +139,7 @@ void main() {
     String selector(List<String> receivedProtocols) => passedProtocol;
 
     server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
     server.listen((HttpRequest request) {
       expect(
         WebSocketTransformer.upgrade(request, protocolSelector: selector),
@@ -133,6 +151,7 @@ void main() {
       'ws://localhost:${server.port}',
       protocols: [failedProtocol],
     );
+    expect(channel.ready, throwsA(isA<WebSocketException>()));
     expect(
       channel.stream.drain(),
       throwsA(isA<WebSocketChannelException>()),
@@ -144,6 +163,7 @@ void main() {
     String selector(List<String> receivedProtocols) => passedProtocol;
 
     server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
     server.listen((HttpRequest request) async {
       final webSocket = await WebSocketTransformer.upgrade(
         request,
@@ -155,7 +175,58 @@ void main() {
 
     final channel = IOWebSocketChannel.connect('ws://localhost:${server.port}',
         protocols: [passedProtocol]);
+
+    expect(channel.ready, completes);
+
     await channel.stream.drain();
     expect(channel.protocol, passedProtocol);
+  });
+
+  test('.connects with a timeout parameters specified', () async {
+    server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
+    server.transform(WebSocketTransformer()).listen((webSocket) {
+      expect(() async {
+        final channel = IOWebSocketChannel(webSocket);
+        await channel.stream.drain();
+        expect(channel.closeCode, equals(5678));
+        expect(channel.closeReason, equals('raisin'));
+      }(), completes);
+    });
+
+    final channel = IOWebSocketChannel.connect(
+      'ws://localhost:${server.port}',
+      connectTimeout: const Duration(milliseconds: 1000),
+    );
+    expect(channel.ready, completes);
+    await channel.sink.close(5678, 'raisin');
+  });
+
+  test('.respects timeout parameter when trying to connect', () async {
+    server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
+    server
+        .transform(StreamTransformer<HttpRequest, HttpRequest>.fromHandlers(
+            handleData: (data, sink) {
+          // Wait before we handle this request, to give the timeout a chance to
+          // kick in. We still want to make sure that we handle the request
+          // afterwards to not have false positives with the timeout
+          Timer(const Duration(milliseconds: 800), () {
+            sink.add(data);
+          });
+        }))
+        .transform(WebSocketTransformer())
+        .listen((webSocket) {
+          final channel = IOWebSocketChannel(webSocket);
+          channel.stream.drain();
+        });
+
+    final channel = IOWebSocketChannel.connect(
+      'ws://localhost:${server.port}',
+      connectTimeout: const Duration(milliseconds: 500),
+    );
+
+    expect(channel.ready, throwsA(isA<TimeoutException>()));
+    expect(channel.stream.drain(), throwsA(isA<WebSocketChannelException>()));
   });
 }