+++ pkgs/web_socket_channel/CHANGELOG.md
@@ -1,5 +1,9 @@
-## 2.2.1-dev
+## 2.3.0
 
+- Added a Future `ready` property to `WebSocketChannel`, which completes when
+  the connection is established
+- Added a `connectTimeout` parameter to the `IOWebSocketChannel.connect` factory,
+  which controls the timeout of the WebSocket Future.
 - Use platform agnostic code in README example.
 
 ## 2.2.0